+++ amappy/amappy/__init__.py
@@ -1,0 +1,10 @@
+"""
+Amappy:
+"""
+
+__version__ = '0.0.0'
+__author__ = "Julien Vitard"
+__author_email__ = "julienvitard+amappy@gmail.com"
+__url__ = "https://github.com/julienvitard/amappy"
+
+VERSION = tuple(__version__.split('.'))