+++ CHANGES.rst
@@ -2,4 +2,4 @@ Changes
 =======
 
 2015-05-14: 0.0.0
-    * Creation of **Amapy** project
+    * Creation of **amappy** project